+++ CHANGELOG.md
@@ -5,7 +5,7 @@
 ### Features
 
 - Include app permissions with event (#1984)
-- Add locale context to event (#2036)
+- Add culture context to event (#2036)
 
 ## 7.23.0